make reusable transaction badge component
